+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Summary:
     complicated anyway (warn added to docs).
 -   Changed `--force` to be the same as `--defaults --overwrite`.
 -   Copied files will reflect permissions on the same files in the template.
+-   Copier now uses `git clone --filter=blob:none` when cloning, to be faster.
 
 ### Deprecated